+++ apps/example-expo/features/example-bottom-sheet-with-keyboard/managed-text-input.tsx
@@ -1,0 +1,25 @@
+import { forwardRef } from 'react'
+import { TextInputProps } from 'react-native'
+import { TextInput } from 'react-native-gesture-handler'
+import { useInputFocus } from 'react-native-the-sheet'
+
+export const ManagedTextInput = forwardRef<TextInput, TextInputProps>(
+  ({ onFocus: propOnFocus, onBlur: propOnBlur, ...rest }, ref) => {
+    const { onFocus, onBlur } = useInputFocus()
+
+    return (
+      <TextInput
+        ref={ref}
+        {...rest}
+        onFocus={(e) => {
+          onFocus()
+          propOnFocus?.(e)
+        }}
+        onBlur={(e) => {
+          onBlur()
+          propOnBlur?.(e)
+        }}
+      />
+    )
+  },
+)