+++ packages/react-native-the-sheet/src/core/height-budget/types.ts
@@ -4,12 +4,12 @@ import type Animated from 'react-native-reanimated'
 import type { SharedValue } from 'react-native-reanimated'
 
 export type HeightBudgetContextType = {
-  maxHeight: number
+  maxHeight: SharedValue<number>
   staticHeights: SharedValue<Record<string, number>>
 }
 
 export type HeightBudgetProviderProps = PropsWithChildren & {
-  maxHeight: number
+  maxHeight: SharedValue<number>
 }
 
 type AnimatedViewProps = ComponentProps<typeof Animated.View>