+++ packages/react-native-the-sheet/src/core/input-focus-provider/index.ts
@@ -1,0 +1,2 @@
+export * from './input-focus-provider'
+export * from './types'